can't do one file system on macos

+++ base/.zshrc
@@ -44,7 +44,10 @@ alias gca='git commit -v -a'
 alias gcam='git commit -a -m'
 alias s='sudo zsh'
 alias ta='tmux attach'
-alias rm='rm -i --one-file-system'
+
+if [[ "$(uname)" != "Darwin" ]]; then
+    alias rm='rm -i --one-file-system'
+fi
 
 # per-host customizations
 if [[ $SHORT_HOST == dev ]]; then
